[Radoub.Formats] fix: KEY/BIF integer overflow in offset and size arithmetic (#2244)
KeyReader.ReadBifEntries / ReadKeyEntries computed
`offset + (i * EntrySize)` in uint then cast to int. With a header
claiming a near-uint-max table offset, the cast yields a negative int
that passes the bounds check (`entryOffset + EntrySize > buffer.Length`
underflows back into a small positive value) and causes the next
BitConverter call to throw ArgumentOutOfRangeException instead of a
diagnosable InvalidDataException.

Fix: match ErfReader.ReadResources — long-promoted arithmetic with an
explicit `> int.MaxValue` guard, then cast. Same pattern applied to the
BIF filename-table lookup so a malicious FilenameOffset can't wrap.

BifFixedResource.TotalSize was `PartCount * PartSize` with both uint;
2^16 * 2^16 wraps silently to 0. Promoted to ulong. The single consumer
(BifReaderTests:59) updated to 512UL.

Tests:
- KeyReader_BifEntryOffsetOverflow_Throws, KeyReader_KeyEntryOffsetOverflow_Throws,
  BifFixedResource_TotalSize_DoesNotWrap now pass.
- KeyReaderTests + BifReaderTests: 22/22 green (no regression).

+++ Radoub.Formats/Radoub.Formats/Bif/BifFile.cs
@@ -148,6 +148,7 @@ public class BifFixedResource
 
     /// <summary>
     /// Total size of the resource (PartCount * PartSize).
+    /// Returned as ulong because uint * uint silently wraps at 2^32 (#2244).
     /// </summary>
-    public uint TotalSize => PartCount * PartSize;
+    public ulong TotalSize => (ulong)PartCount * PartSize;
 }
